@@ -1,61 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-	<UserSecretsId>89325659-56d0-436e-993a-e44aa19ed5e7</UserSecretsId>
-    <Platforms>AnyCPU;x64</Platforms>
-    <Authors>Nikolay Medvedev</Authors>
-    <Company>Nick219</Company>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <UserSecretsId>89325659-56d0-436e-993a-e44aa19ed5e7</UserSecretsId>
+        <Platforms>AnyCPU;x64</Platforms>
+        <Authors>Nikolay Medvedev</Authors>
+        <Company>Nick219</Company>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <Compile Remove="ClientApp\samples\**" />
-    <Compile Remove="Thrash\**" />
-    <Content Remove="ClientApp\samples\**" />
-    <Content Remove="Thrash\**" />
-    <EmbeddedResource Remove="ClientApp\samples\**" />
-    <EmbeddedResource Remove="Thrash\**" />
-    <None Remove="ClientApp\samples\**" />
-    <None Remove="Thrash\**" />
-    <TypeScriptCompile Remove="ClientApp\samples\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Remove="wwwroot\js\menu.tsx" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="ClientApp\public\bootstrap.css" />
-    <None Remove="ClientApp\public\bootstrap.css.map" />
-    <None Remove="ClientApp\public\bootstrap.min.css" />
-    <None Remove="ClientApp\public\bootstrap.min.css.map" />
-    <None Remove="ClientApp\public\react.css" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="JavaScriptEngineSwitcher.ChakraCore" Version="3.9.1" />
-    <PackageReference Include="JavaScriptEngineSwitcher.ChakraCore.Native.linux-x64" Version="3.9.1" />
-    <PackageReference Include="JavaScriptEngineSwitcher.ChakraCore.Native.osx-x64" Version="3.9.1" />
-    <PackageReference Include="JavaScriptEngineSwitcher.ChakraCore.Native.win-x64" Version="3.9.1" />
-    <PackageReference Include="JavaScriptEngineSwitcher.ChakraCore.Native.win-x86" Version="3.9.1" />
-    <PackageReference Include="JavaScriptEngineSwitcher.Extensions.MsDependencyInjection" Version="3.3.0" />
-    <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.1.16" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.8">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.8" />
-    <PackageReference Include="Microsoft.TypeScript.MSBuild" Version="4.4.2">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="5.0.1" />
-    <PackageReference Include="React.AspNet" Version="5.2.11" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.1.4" />
-  </ItemGroup>
-
-  <Target Name="PreBuild" BeforeTargets="PreBuildEvent">
-    <Exec Command="cd $(SpaRoot)ClientApp\node_modules\ || cd $(SpaRoot)ClientApp\ &amp;&amp; npm install" />
-  </Target>
+    <ItemGroup>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.8" />
+        <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="5.0.1" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.1.4" />
+    </ItemGroup>
 
 </Project>
