@@ -1,59 +1,79 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using RandomSongSearchEngine.Dto;
 using RandomSongSearchEngine.Models;
 using System;
 using System.Threading.Tasks;
+using RandomSongSearchEngine.Data.DTO;
 
-namespace RandomSongSearchEngine.Controllers
+namespace RandomSongSearchEngine.Controllers;
+
+// [Authorize]
+[ApiController]
+[Route("api/read")]
+
+public class ReadController : ControllerBase
 {
-    //[Authorize]
-    //[DisableCors]
-    //[EnableCors]
-    [ApiController]
-    [Route("api/read")]
+    private readonly ILogger<ReadController> _logger;
+    private readonly IServiceScopeFactory _serviceScopeFactory;
 
-    public class ReadController : ControllerBase
+    public ReadController(IServiceScopeFactory serviceScopeFactory, ILogger<ReadController> logger)
     {
-        private readonly ILogger<ReadController> _logger;
-        private readonly IServiceScopeFactory _serviceScopeFactory;
- 
-        public ReadController(IServiceScopeFactory serviceScopeFactory, ILogger<ReadController> logger)
-        {
-            _serviceScopeFactory = serviceScopeFactory;
-            _logger = logger;
-            //_logger = _serviceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<ILogger<ReadController>>();
-        }
+        _serviceScopeFactory = serviceScopeFactory;
+        _logger = logger;
+        // _logger = _serviceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<ILogger<ReadController>>();
+    }
 
-        [HttpGet]
-        public async Task<ActionResult<SongDto>> OnGetGenreListAsync()
-        {
-            try
-            {
-                using var scope = _serviceScopeFactory.CreateScope();
-                return await new ReadModel(scope).ReadGenreListAsync();
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "[ReadController: OnGet Error]");
-                return new SongDto() { ErrorMessageResponse = "[ReadController: OnGet Error]" };
-            }
-        }
+    [HttpGet]
+    public async Task<ActionResult<SongDto>> OnGetGenreListAsync()
+    {
+        // CORS ручная настройка
+        // credentials: "include" или "same-origin"
+        // HttpContext.Response.Headers.Add("Access-Control-Allow-Origin", "http://localhost:3000");
+        // HttpContext.Response.Headers.Add("Access-Control-Allow-Credentials", "true");
 
-        [HttpPost]
-        public async Task<ActionResult<SongDto>> GetRandomSongAsync([FromBody] SongDto dto)
+        try
         {
-            try
-            {
-                using var scope = _serviceScopeFactory.CreateScope();
-                return await new ReadModel(scope).ReadRandomSongAsync(dto);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "[ReadController: OnPost Error]");
-                return new SongDto() { ErrorMessageResponse = "[ReadController: OnPost Error]" };
-            }
+            using var scope = _serviceScopeFactory.CreateScope();
+            return await new ReadModel(scope).ReadGenreListAsync();
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "[ReadController: OnGet Error]");
+            return new SongDto() {ErrorMessageResponse = "[ReadController: OnGet Error]"};
         }
     }
+
+    [HttpPost]
+    public async Task<ActionResult<SongDto>> GetRandomSongAsync([FromBody] SongDto dto)
+    {
+        // CORS ручная настройка
+        // HttpContext.Response.Headers.Add("Access-Control-Allow-Origin", "http://localhost:3000");
+        // HttpContext.Response.Headers.Add("Access-Control-Allow-Credentials", "true");
+        // HttpContext.Response.Headers.Add("Access-Control-Allow-Methods", "POST");
+
+        try
+        {
+            using var scope = _serviceScopeFactory.CreateScope();
+            return await new ReadModel(scope).ReadRandomSongAsync(dto);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "[ReadController: OnPost Error]");
+            return new SongDto() {ErrorMessageResponse = "[ReadController: OnPost Error]"};
+        }
+    }
+
+    // CORS ручная настройка
+    /*
+    [HttpOptions]
+    public ActionResult Options()
+    {
+        HttpContext.Response.Headers.Add("Access-Control-Allow-Origin", "http://localhost:3000");
+        HttpContext.Response.Headers.Add("Access-Control-Allow-Credentials", "true");
+        HttpContext.Response.Headers.Add("Access-Control-Allow-Methods", "OPTIONS");
+        HttpContext.Response.Headers.Add("Access-Control-Allow-Headers", "Content-type");
+        return Ok();
+    }
+    */
 }
