@@ -1,9 +1,8 @@
-﻿namespace RandomSongSearchEngine.Data
+﻿namespace RandomSongSearchEngine.Data;
+
+public class UserEntity
 {
-    public class UserEntity
-    {
-        public int Id { get; set; }
-        public string Email { get; set; }
-        public string Password { get; set; }
-    }
+    public int Id { get; set; }
+    public string Email { get; set; }
+    public string Password { get; set; }
 }
