@@ -1,13 +1,12 @@
 ﻿using Microsoft.Extensions.Logging;
 
-namespace RandomSongSearchEngine.Services.Logger
+namespace RandomSongSearchEngine.Services.Logger;
+
+public static class FileLoggerExtensions
 {
-    public static class FileLoggerExtensions
+    public static ILoggerFactory AddFile(this ILoggerFactory factory, string filePath)
     {
-        public static ILoggerFactory AddFile(this ILoggerFactory factory, string filePath)
-        {
-            factory.AddProvider(new FileLoggerProvider(filePath));
-            return factory;
-        }
+        factory.AddProvider(new FileLoggerProvider(filePath));
+        return factory;
     }
 }
