@@ -1,18 +1,17 @@
 ﻿using System;
 
-namespace RandomSongSearchEngine.Services.Services
-{
-    /// <summary>
-    /// Моё исключение
-    /// </summary>
-    public class DataExistsException : Exception
-    {
-        public DataExistsException(string message) : base(message)
-        {
-        }
+namespace RandomSongSearchEngine.Services.Services;
 
-        public DataExistsException(string message, Exception inner) : base(message, inner)
-        {
-        }
+/// <summary>
+/// Моё исключение
+/// </summary>
+public class DataExistsException : Exception
+{
+    public DataExistsException(string message) : base(message)
+    {
+    }
+
+    public DataExistsException(string message, Exception inner) : base(message, inner)
+    {
     }
 }
