@@ -3,34 +3,34 @@ using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
 using RandomSongSearchEngine.Controllers;
-using RandomSongSearchEngine.Dto;
 using RandomSongSearchEngine.Models;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using RandomSongSearchEngine.Data.DTO;
 
 namespace MSTest
 {
     [TestClass]
     public class ReadTest
     {
-        const int GenresCount = 44;
-        IServiceScope fakeScope;
-        ReadModel readModel;
+        private const int GenresCount = 44;
+        private IServiceScope _fakeScope;
+        private ReadModel _readModel;
 
         [TestInitialize]
         public void Initialize()
         {
             FakeLoggerErrors.ExceptionMessage = "";
             FakeLoggerErrors.LogErrorMessage = "";
-            fakeScope = new FakeScope<ReadModel>().ServiceScope;
-            readModel = new ReadModel(fakeScope);
+            _fakeScope = new FakeScope<ReadModel>().ServiceScope;
+            _readModel = new ReadModel(_fakeScope);
         }
 
         [TestMethod]
         public async Task ShouldBe44GenresTest()
         {
-            var response = await readModel.ReadGenreListAsync();
+            var response = await _readModel.ReadGenreListAsync();
             Assert.AreEqual(GenresCount, response.GenreListResponse.Count);
         }
 
@@ -38,7 +38,7 @@ namespace MSTest
         public async Task ShouldReadRandomSongTest()
         {
             var request = new SongDto() { SongGenres = new List<int>() { 2 } };
-            var response = await readModel.ReadRandomSongAsync(request);
+            var response = await _readModel.ReadRandomSongAsync(request);
             Assert.AreEqual("test title", response.TitleResponse);
         }
 
@@ -46,21 +46,21 @@ namespace MSTest
         public async Task ShouldResponseEmpltyTitleIfWtfTest()
         {
             var frontRequest = new SongDto() { SongGenres = new List<int>() { 1000 } };
-            var result = await readModel.ReadRandomSongAsync(frontRequest);
+            var result = await _readModel.ReadRandomSongAsync(frontRequest);
             Assert.AreEqual("", result.TitleResponse);
         }
 
         [TestMethod]
         public async Task IfNullShouldLoggingErrorInsideModelTest()
         {
-            _ = await readModel.ReadRandomSongAsync(null);
+            _ = await _readModel.ReadRandomSongAsync(null);
             Assert.AreNotEqual("[IndexModel: OnPost Error]", FakeLoggerErrors.LogErrorMessage);
         }
 
         [TestMethod]
         public async Task IfNullShouldResponseEmptyTitleTest()
         {
-            var response = await readModel.ReadRandomSongAsync(null);
+            var response = await _readModel.ReadRandomSongAsync(null);
             Assert.AreEqual("", response.TitleResponse);
         }
 
@@ -82,7 +82,7 @@ namespace MSTest
         {
             var mockLogger = Substitute.For<ILogger<ReadController>>();
             var fakeServiceScopeFactory = Substitute.For<IServiceScopeFactory>();
-            fakeServiceScopeFactory.CreateScope().Returns(fakeScope);
+            fakeServiceScopeFactory.CreateScope().Returns(_fakeScope);
             var readController = new ReadController(fakeServiceScopeFactory, mockLogger);
 
             var response = (await readController.GetRandomSongAsync(null)).Value;//.Result.Value;
@@ -93,7 +93,7 @@ namespace MSTest
         [TestCleanup]
         public void TestCleanup()
         {
-            fakeScope.Dispose();
+            _fakeScope.Dispose();
         }
     }
 }
